+++ learning_resources_search/views.py
@@ -1,12 +1,10 @@
 """View for search"""
 
 import logging
-from functools import wraps
 from itertools import chain
 
 from django.conf import settings
 from django.utils.decorators import method_decorator
-from django.views.decorators.cache import cache_page
 from drf_spectacular.utils import OpenApiParameter, extend_schema, extend_schema_view
 from opensearchpy.exceptions import TransportError
 from rest_framework import mixins, viewsets
@@ -33,23 +31,9 @@ from learning_resources_search.serializers import (
     PercolateQuerySerializer,
     PercolateQuerySubscriptionRequestSerializer,
 )
+from main.utils import cache_page_for_anonymous_users
 
 log = logging.getLogger(__name__)
-
-
-def cache_page_for_anonymous_users(*cache_args, **cache_kwargs):
-    def inner_decorator(func):
-        @wraps(func)
-        def inner_function(request, *args, **kwargs):
-            if not request.user.is_authenticated:
-                return cache_page(*cache_args, **cache_kwargs)(func)(
-                    request, *args, **kwargs
-                )
-            return func(request, *args, **kwargs)
-
-        return inner_function
-
-    return inner_decorator
 
 
 class ESView(APIView):
